refactor: move the image protocols into graphics/, with regis/ beneath it
The protocols that produce an Image: Sixel (DCS), kitty graphics (APC), Contour's
own upload protocol (MessageParser), and ReGIS -- which keeps a directory of its
own, one level further down.

Image itself is NOT here; it went to core/ two commits ago, because ColorPalette
includes it and 22 files across three modules name it. That is what makes this
directory a sharp thing rather than a vague one: graphics/ is the DECODERS, and
what they all decode INTO is shared vocabulary a layer below.

regis/ nests rather than sitting beside them because it is one protocol among
four, not a peer concern of "grid" and "input" -- which is what a top-level
regis/ read as. Its rasterizer already includes Image, so it is an image producer
exactly as the other three are; it merely needs seven translation units where they
need one each. The shape is vthost/{client,imsg,proto,tmux}: a concern with a
sub-protocol big enough to want its own directory.

Three tests that had no source of their own get a home by what they exercise
rather than by what they were named after:

  GoodImageProtocol_test  drives MessageParser and Image through Screen.
  XtSmGraphics_test       is sixel and ReGIS geometry and colour registers
                          throughout; its [screen] tag names the harness, not
                          the subject.
  ReGIS_test              is the end-to-end suite, so it becomes
                          regis/ReGISScreen_test.cpp -- it was reading as a stray
                          peer of ReGISParser_test.cpp, which it is not.
