+++ src/vtbackend/graphics/SixelParser.hpp
