+++ src/vtbackend/graphics/regis/ReGISColor_test.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/regis/ReGISColor.hpp>
+#include <vtbackend/graphics/regis/ReGISColor.hpp>
 
 #include <catch2/catch_test_macros.hpp>
