+++ src/vtbackend/graphics/regis/ReGISFont_test.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/regis/ReGISFont.hpp>
+#include <vtbackend/graphics/regis/ReGISFont.hpp>
 
 #include <catch2/catch_test_macros.hpp>
