+++ src/vtbackend/graphics/regis/ReGISRasterizer_test.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/regis/ReGISRasterizer.hpp>
+#include <vtbackend/graphics/regis/ReGISRasterizer.hpp>
 
 #include <catch2/catch_test_macros.hpp>
