+++ src/vtbackend/graphics/regis/ReGISTextRasterizer.cpp
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/regis/ReGISTextRasterizer.hpp>
+#include <vtbackend/graphics/regis/ReGISTextRasterizer.hpp>
 
 #include <algorithm>
 #include <cmath>
